Move exception from server to spring-cloud-skipper
- As these exceptions are not only server related, move
  to common shared package so that we can use same
  classes in a client as well. This would be to
  i.e. propagate ReleaseNotFoundException from server
  to client via rest.
- Made some static methods in ReleaseNotFoundException
  private as if those are only accessed in tests then
  its better to hardcode asserts in tests.
- Bit of polish and javadocs.
- Fixes #146
